requirements.txt
# Core dependencies
ansible>=8.0.0
ansible-core>=2.15.0
click>=8.0.0
flask>=2.3.0
flask-sqlalchemy>=3.0.0
sqlalchemy>=2.0.0
pyyaml>=6.0
jinja2>=3.1.0
cryptography>=41.0.0
paramiko>=3.0.0

# Database
sqlite3

# Utilities
requests>=2.31.0
python-dateutil>=2.8.0
tabulate>=0.9.0
colorama>=0.4.0

# Development dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Optional monitoring
psutil>=5.9.0
